ai-test-generator/pyproject.toml
[project]
name = "ai-test-generator"
version = "1.0.0"
description = "AI-powered test case generator from OpenAPI/Swagger specs using Claude API"
requires-python = ">=3.11"
authors = [{name = "Abhinav Gautam"}]
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
testgen = "src.cli:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
